+++ src/Factory/ServerRequestFactory.php
@@ -3,7 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Factory;
 
-use Innmind\Http\Message\ServerRequest;
+use Innmind\Http\ServerRequest;
 
 /**
  * @psalm-immutable